hotfix: ModifyMyAccountPage -> EditProfilePage로 파일 및 폴더 이름 수정
